172722 - Renamed folder Helpers to Clients AB#172722
